Progress in SIMD brightness

+++ lab04/code/part2/include/weird.h
@@ -1,3 +1,4 @@
 #include "image.h"
+#include <stdint.h>
 
-void weirdimg_transform(struct img_t *image, int brightness_factor);
+void weirdimg_transform(struct img_t *image, int16_t brightness_factor);
